requirements.txt
aniso8601==9.0.1
click==8.0.4
Flask==2.0.3
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
mysqlclient==2.1.0
packaging==21.3
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.32
Werkzeug==2.0.3
